Small DLC GUI touchups
