+++ app/gui/src/main/scala/org/bitcoins/gui/dlc/dialog/AddSigsDLCDialog.scala
@@ -17,7 +17,8 @@ class AddSigsDLCDialog
                                                     Some(file)
                                                   DLCDialog.signFileChosenLabel.text =
                                                     file.toString
-                                                }
+                                                },
+                                              DLCDialog.fileChosenStr -> DLCDialog.signFileChosenLabel
                                             ),
                                             Vector(DLCDialog.dlcSigStr,
                                                    DLCDialog.dlcSignFileStr)) {
